add todo for semver images
