+++ internal/orchestrator/container_updater.go
@@ -5,7 +5,6 @@ import (
 	"slices"
 
 	"github.com/docker/cli/cli/command"
-	semver "go.bug.st/relaxed-semver"
 
 	"github.com/arduino/arduino-app-cli/internal/orchestrator/bricksindex"
 	"github.com/arduino/arduino-app-cli/internal/orchestrator/config"
@@ -67,22 +66,4 @@ func (c *ContainerUpdater) ListUpgradableImages(ctx context.Context) ([]update.U
 	}
 
 	return result, nil
-}
-
-// isNewerVersion returns true if newImage has a strictly higher semver tag than currentImage.
-// Both arguments are expected to share the same image base name.
-// If either tag cannot be parsed as semver, it returns true (treat as upgrade to be safe).
-func isNewerVersion(newImage, currentImage string) bool {
-	_, newTag := parseDockerImage(newImage)
-	_, currentTag := parseDockerImage(currentImage)
-
-	newVer, err := semver.Parse(newTag)
-	if err != nil {
-		return true
-	}
-	currentVer, err := semver.Parse(currentTag)
-	if err != nil {
-		return true
-	}
-	return currentVer.LessThan(newVer)
 }
